Moves a file to a different package
